requirements.txt
selenium
pytest
python-dotenv
pytest-html
allure-pytest
pytest-check
